stock_analysis_ai/requirements.txt
yfinance
pandas
numpy
streamlit
streamlit
plotly
rich
pyyaml
